tire logging

+++ persistence/src/main/java/cz/fi/muni/pa165/dao/TireDaoImpl.java
@@ -1,6 +1,9 @@
 package cz.fi.muni.pa165.dao;
 
 import cz.fi.muni.pa165.entity.Tire;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
@@ -14,11 +17,17 @@ import javax.persistence.TypedQuery;
 @Repository
 public class TireDaoImpl implements TireDao {
 
+    private final static Logger log = LoggerFactory.getLogger(CarDaoImpl.class);
+
     @PersistenceContext
     private EntityManager em;
 
     @Override
     public void create(Tire tire) {
+        if (tire == null){
+            throw new DataAccessException("Attempting to create null Tire entity"){};
+        }
+        log.debug("DAO - create Tire");
         em.persist(tire);
     }
 
@@ -29,6 +38,7 @@ public class TireDaoImpl implements TireDao {
 
     @Override
     public Tire findById(Long id) {
+        log.debug("DAO - Find all tires");
         return em.find(Tire.class, id);
     }
 
@@ -39,16 +49,25 @@ public class TireDaoImpl implements TireDao {
                 Tire.class);
 
         query.setParameter("manufacturer", manufacturer);
+        log.debug("DAO - Find tire by MANUFACTURER {}", manufacturer);
         return query.getResultList();
     }
 
     @Override
     public void remove(Tire tire) {
+        if (tire == null){
+            throw new DataAccessException("Attempting to remove null Tire."){};
+        }
+        log.debug("DAO - removing tire with ID {}", tire.getId());
         em.remove(tire);
     }
 
     @Override
     public void update (Tire tire) {
+        if (tire == null){
+            throw new DataAccessException("Attempting to update null Tire."){};
+        }
+        log.debug("DAO - updating tire with ID {}", tire.getId());
         em.merge(tire);
     }
 }
